refactor(invoices): moves the unpaid invoices table SP to CTE
This commit remove the UnbalancedInvoicePaymentsTable() stored procedure.
Instead, we use a common table expression called directly from the report
code.  This increases the clarity of the code and decreases the amount
of I/O needed.  Unfortunately, due to the API of the PIVOT() function,
we still need to create a temporary table.