chore: refactor stored procedures
This commit refactors the stored procedures (SP) to remove
redudent procedures (or unused SPs).  It also rewrite the stock_movement_status
stored procedures to use common table expressions CTEs rather than the the
temporary table mess that it used to be.  This should hopefully improve
the performance of our stock movement tables a bit.